even in Haskell I need to learn how to be able to safely use the
constructor of a data type (enum in rust?) as the type of a parameter in
a function. Doing pattern matching when you know that it can only be
what you passed in is annoying. On a team, other people might not know
that so it needs to encoded into the type system somehow "this parameter
must specifically be a RedCount, not a Count"